services/shared/pyproject.toml
[project]
name = "aegis-shared"
version = "0.1.0"
description = "Shared library for Aegis microservices"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "structlog>=24.4.0",
    "google-cloud-pubsub>=2.23.0",
    "google-cloud-firestore>=2.18.0",
    "google-cloud-storage>=2.18.0",
    "google-cloud-logging>=3.10.0",
    "google-cloud-aiplatform>=1.70.0",
    "google-genai>=0.3.0",
    "firebase-admin>=6.5.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["aegis_shared*"]
